refactor(settings): extract cards pane into eh-settings-cards.js
The Settings view was a 1029-line monolith. This commit pulls the
per-card master enable/disable list (filter, count summary, demo
gating, toggle) out into its own component so the upcoming tabbed
shell can mount it as a tab without dragging the rest of Settings
along.

eh-settings-view stays as the host: HAE panel, hidden discoveries,
and now <eh-settings-cards>. No behaviour change.

The reorder-cards button (currently in the nav dropdown) is wired
into the cards pane in a later commit; for now both entry points
work.

Refs #383.
